fix(spellcheck): render spellcheck in a clone
test a different rendering strategy, instead of updating the text
with markup for hiliting create a clone, do the hiliting there and
render it below the actual text. advantage is that there is no
cursor manipulation involved, it doesn't change markup so it doesn't
need to store and restore cursor position.

known issues:
- find and replace - before it was using hilite node to replace the
  word, it will require some changes to manipulate the text instead
- corrections dropdown - because the errors are rendered below the text,
  there would be no click event on this.