fix spellcheck for editor2 (#305)
* fix(spellcheck): render spellcheck in a clone

test a different rendering strategy, instead of updating the text
with markup for hiliting create a clone, do the hiliting there and
render it below the actual text. advantage is that there is no
cursor manipulation involved, it doesn't change markup so it doesn't
need to store and restore cursor position.

known issues:
- find and replace - before it was using hilite node to replace the
  word, it will require some changes to manipulate the text instead
- corrections dropdown - because the errors are rendered below the text,
  there would be no click event on this.

* test detecting error on right click

* fix spellcheck suggestions dropdown

* improve htmlCollections handling

in first case there can be only 1 or 0 items, so no need for the loop.

in second case cache the length, using array slice was slower so keep the
htmlcollection in the loop.