set corresp anpa categ when subservice required
Sets the anpa category corresponding to the required subservice:
if a subservice field (defined in vocabularies) was declared as
required in a content profilebthen make sure that the corresponding
anpa category was added to the anpa_category field in the newly
created item. Without this value the anpa category field is not
displayed.