fix(editor): fix highlights alignment
there was font/line size set for spans within editor to match
it's defined size which avoids inserting that by chrome.
but size was defined on `p` tag and not on `pre` tag so then
it would be different for preformatted text and spans used
for spellcheck highlights would have different size than the text.

SD-5155