fix(authoring): fix reverting to an empty item after save
populate item with default values when editing starts, so that it
can be reverted to initial state, even if origItem changed after saving.

so this is actually a workaround for js inheritance, ideally we wouldn't
use it then, but this seems to be quite some work due to those `_locked`
and `_editable` and other flags.