fix(editor3): documented attributes and added find-replace-target
When the Editor3 directive is being used, it registers itself with a
service that allows external components to run find & replace operations
on it. Only one editor can be registered at a time, and this flag
enables that operation.

The `data-find-replace-target` attribute registers the editor with the
find & replace service and de-registers it when its scope is destroyed.
Only one editor can be registered at a time.