fix(editor3): enable multiple editors (#1522)
* fix(editor3): documented attributes and added find-replace-target

When the Editor3 directive is being used, it registers itself with a
service that allows external components to run find & replace operations
on it. Only one editor can be registered at a time, and this flag
enables that operation.

The `data-find-replace-target` attribute registers the editor with the
find & replace service and de-registers it when its scope is destroyed.
Only one editor can be registered at a time.

* feat(editor3): allow editor3 services to override editor2

- Added missing interface methods to the editor3 service, so that it can
replace editor2 when needed. Assures SendItem & Macros work correctly.

- Added config `features.onlyEditor3` which completely disables editor2.

- Added single-line option to the editor, enabled by adding
  `data-single-line` as an attribute (no value needed).

- Added editor3 version for all fields on items.

- Fixed problem with hovering toolbar, so that when an editor goes out
  of view, the toolbar goes away (instead of staying sticky when no
editor is in view anymore).

- Added several CSS fixes to acommodate editor3.

- Made formatting options for special content work. Editor formats
  supported: picture, anchor, embed, table

* feat(editor3): new set of icons

Updates the new set of icons provided by the design team.

* feat(editor3): added editorResolver service