ui(editor3/embed): allow adding embed codes
This is a highly insecure implementation which allows pasting random
HTML into the embed input, which alternatively gets injected into the
editor's content.

It's a temporary solution after discussing with @karelpetrak. Assuming
this software is used only internally and that potential employee X with
malicious intentions can find worse ways to harm the company, it should
be ok for now.

The 'correct' implementation would be to sanitize the input and see if
it's a valid embed code for any of the supported providers, which would
prove a lot more work and would be harder to maintain due to the fact
that embed codes might change all the time. Alternatively, we could also
simply santitize the URLs presented in the HTML, making sure they
represent trusted sources.

Change-Id: I63fe77e32b69038a86b0e7c927cd8b9728295df5