[AUTHORING] handle "authors" field
This commit add "authors" field using a drop-down list similar to
subjects.

A new "selectEntireCategory" variable has been added to MetaTermsDirective to indicate if the parent category can be selected as a whole, or if a subcategory is needed.

A new sub_label can be used in terms. If present, the display of the
term show the term first in a label, then the sub_label.

SDESK-1703