core/editor3: consolidated toolbar buttons
Removed CommentButton, LinkButton and created SelectionButton which can
be added to the toolbar and customized. It is only enabled when a
selection is active in the editor. This is to avoid repetitive code for
CommentButton, LinkButton, AnnotationButton, etc.

Additionally introduced IconButton to replace LinkButton and ImageButton.

Renamed {Block,Inline}StyleControls to {Block,Inline}StyleButtons for
consistency.

Change-Id: I80ead1e72712032953cdad4aa41b8837222fa594