core/editor3: refactor toolbar popup and move actions to their own
Previously popup actions were created in the generic ToolbarPopup
component. This change moves each action into its own popup. It also
decouples the functions that display and hide popups from the toolbar
and moves it into the Redux flow as an action/reducer and a store prop
called 'popup'.

Additionally, SelectionButton now creates the data as an object having a
key 'selection', as opposed to only the selection. This is to
accommodate sending other data along.

Change-Id: I1128e08a9a07aed58eb93991a21839456d4571fb