[AUTHORING] [METADATA]: raise limit of authors to 200
authors fetching was using the default max_results (i.e. 25), resulting
in trouble when a instance has more authors.
This patch is a quick fix to raise this limit to the maximum (200)
before checking a more clean solution which need discussion (should we
fetch all the pages, or use pagination?).

SDFID-195