core/editor3: better accommodate longer texts in highlights popup
Depending on the editor's scroll position, at times the popup would be
cut off at the bottom, making it impossible to see the entire content.
This change fixes the problem by introducting a maximum height to the
popup and a scroll bar. It also introduces a limit as to how far down
the popup is displayed.

A similar approach is applied to the annotation editor to allow for
longer texts by making the editor have a maximum height and an automatic
scrollbar.

Fixes SDESK-2184

Change-Id: I02460d367edceb0aa029c2980bf547d4065ef919