[AUTHORING] [WORD COUNT] synchronized word count algorithm with backend one
Word count should now be the same in backend and client, making numbers
coherent in differents places where it's used. In some cases the numbers
may still differ because backend and client ways to clean HTML are not
the same, but in most cases the current behaviour should be alright.

A new test is using cases handled by the regex, the same test is done
in backend.

SDFID-194