[INGEST] disabled autocompletion on password for feeding service
autocomplete="off" on <input> is not respected anymore by modern
browsers, and autocomplete="off" on <form> is not working either.
Furthermore, Chromium respect autocomplete="new-password" but Firefox
is not (as of 59.0.2).
A hack is used for Firefox with dummy login/password field, which
disabled autocompletion, as it seems to be the only working way (beside
setting password with code).
For Chromium, autocomplete="new-password" should be used for each new
password input.

SDESK-2745