[SDESK-2805] Fix bug when toggling suggested style
When you suggest an inline style change, if you toggle it back in
suggestion mode would remove the suggestion but not the style, so
you would have to toggle it back again but that would end up in a
new suggestion to toggle the style off.

This commit fixes that bug, that was actually created by modifying the
editor selection when handling the suggestion highlight.