Display commented text on unresolved comments inside widget
We don't save commented text on each comment before they are resolved
(as the highlighted text can change in the editor) so these comments
were lacking the "commented text" property in the view.

Now we search inside the editor content for highlighted text for each
comment. It was a bit tricky to do as we sometimes have orphan comments
(their text was deleted but they were not) that could appear on the
sidebar widget but not on the editor.

SDFID-441