[SETTINGS][AUTHORING][VOCABULARIES] use popup_width in categories + hide field in settings
popup_width was not working in settings because the popup is not created
in the same way as for new controlled vocabularies. This patch fixes it.

In settings, the popup_width field was visible for all vocabularies, and it
was not making sense for most of them. This is not the case anymore,
service is used to check if it's a new CV, and the field is not
displayed otherwise, except for categories.

SDANSA-172