Mantain cursor position on table cell when undoing changes (#2762)
* Mantain cursor position on table cell when undoing changes

SDESK-3603

Each cell is a draft-js instance, so each cell has its own editor state
that we save on the table block data. In order to do this we have to
serialize the state but this doesn't save selection state, which means
it's not saved on state changes (therefore it's not on the "undo stack")

We do save the active cell selection state on the global state to be
able to write in the cell normally, but we don't keep a history of
selections

When pressing Ctrl+Z the selection (caret) mantained the same end position,
which means it would overflow if we were undoing one or more characters.

I tried several approaches and the only one I think it works (without
too much code like a selection history or saving the selection state in
the block data) is the following: Just check for overflow whenever we
undo/redo changes on a table. If there's overflow, position the caret at
the end.

* Remove unused method

* Remove focused tests

* Fix tests

getData was changed to use the block key instead of the block object
and that was causing some issues for text because the passed block
was not contained on the passed editor state

* fix lint tests