fix(search): fix list jumping up all the time
it was scrolling on top when it got results, but that scrolling
to top triggers another refresh, and that triggered another update,
and so on.

now it will only scroll on top when refresh btn is clicked and
starts the query later, so scroll handler won't trigger another load.

SDESK-4146