fix(search): Fix issue saving changes on search after updating subscriptions
SDESK-4183

Pressing 'Save changes' was causing an error only after user had changed
subscriptions for the current search. The issue was that saving the
subscriptions was changing the `etag` in the server but the current search
on the client was not being udpated.

As you can see in the code I didn't replace the whole object with the
updated one as the user might lose any changes to the current search
that's being edited, so I just udpated the `etag`.