Force selection when closing popup
SDESK-3885

I put a comment on the code explaining this because it's not intuitive
just by looking at it:

Whenever we hide a popup, the ContentState is not changed so it will
trigger these two DraftJS events:

    * editOnFocus
    * editOnSelect

The first one renders the right selection but the second one uses
global `window.getSelection()` (which doesn't exist, as the editor lost focus)
to check if the editorState selection matches that one. As it doesn't, it renders
the selection in the first character of the first block.
Using `forceSelection` won't trigger those events and the selection will be correct.

> NOTICE

The task is only about links but this also affects other popups like
comments and annotations. If you open one of those popups and then you
click cancel, the error was happening before this fix.