Fix editor2 focus when clicking on placeholder
SDNTB-631

I tested on chrome and edge and they both have the following issue:

When you have a contenteditable element with a `:before` pseudo-element (which is what we use to display the placeholder) then when you click on that placeholder, the `focus` event fires but medium-editor doesn't show the caret nor allows you to write until you focus/click again.

On firefox, it works.

Solution: Preventing mouse events on that pseudo-element.