[INGEST] [SETTINGS] Use Object ID instead of provider name to identity provider
Provider name was used to identify the requesting provider when doing
URL request, but the provider name may not be set yet at this point.
This has been replaced by a generated ID which is compatible with
MongoDB. This way it can be used in the backend in the `_id` field.

SDESK-5342