Fix authors field in monitoring so it does not crash when author _id is not present (#3869)
* Fix authors field in monitoring so it does not crash when author _id is not present.
(in case an item is ingested and we don't have a local user for that author).

* fix e2e test

It started failing seemingly without a reason. The issue is that pressing home on an empty line in editor2 jumpt to the top as if ctrl+home was pressed even if ctrl wasn't pressed. Editor2 is deprecated, so I'm not fixing it.