Fix isssues with _editable state after reverting a version.
When items are previewed from history view, `_editable` is set to false(to prevent editing of old versions). It was not set back to true when performing revert operation and the item was remaining in uneditable state after the revert. There was also another issue that version previewing was called when clicked on "revert" button, because event propagation was not stopped.
The issue was only reproducible on slower network - when simulating fast 4G connection.